:art: Checkbox types
